-added message for no result -built separate fns for submit search
