+++ src/components/SearchTransactions/helpers/parseDate.ts
@@ -1,0 +1,4 @@
+export function parseDate(dateStr: string): Date {
+  const [day, month, year] = dateStr.split("-").map(Number);
+  return new Date(year, month - 1, day);
+}
