+++ src/pages/Expenses/Expenses.tsx
@@ -1,7 +1,12 @@
 import { useEffect, useState } from "react";
 import Expense from "../../components/Expense/Expense";
 import { useInfiniteQuery } from "@tanstack/react-query";
-import { ExpenseResponseItem, Group, UserInfo } from "../../types";
+import {
+  ExpenseParsedFilters,
+  ExpenseResponseItem,
+  Group,
+  UserInfo,
+} from "../../types";
 import { getGroupExpenses } from "../../api/services/api";
 import { useOutletContext } from "react-router-dom";
 import { StyledExpenses } from "./Expenses.styled";
@@ -22,17 +27,19 @@ import {
 } from "../../helpers/getGroupTotalByCurrency";
 import GroupTotalsByCurrencyAnimation from "../../components/Menus/MenuAnimations/GroupTotalsByCurrencyAnimation";
 import Spinner from "../../components/Spinner/Spinner";
+import { FaMagnifyingGlass } from "react-icons/fa6";
 
 const Expenses = () => {
   const selectedExpense = useSignal<ExpenseResponseItem | null>(null);
   const errorMessage = useSignal<string>("");
   const menu = useSignal<string | null>(errorMessage.value ? "error" : null);
-  const { userInfo, group, showBottomBar,expenseParsedFilters } = useOutletContext<{
-    userInfo: UserInfo;
-    group: Group;
-    showBottomBar: Signal<boolean>;
-    expenseParsedFilters:Signal<any>;
-  }>();
+  const { userInfo, group, showBottomBar, expenseParsedFilters } =
+    useOutletContext<{
+      userInfo: UserInfo;
+      group: Group;
+      showBottomBar: Signal<boolean>;
+      expenseParsedFilters: Signal<ExpenseParsedFilters>;
+    }>();
 
   const timeZoneId = userInfo?.timeZone;
   const pageSize = 10;
@@ -53,22 +60,31 @@ const Expenses = () => {
   const shouldOpenMultiCurrencyTable =
     Object.keys(groupTotalsByCurrency).length > 1;
 
-
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage, isFetching } =
     useInfiniteQuery({
-      queryKey: ["groupExpenses", group?.id, pageSize, expenseParsedFilters.value],
+      queryKey: [
+        "groupExpenses",
+        group?.id,
+        pageSize,
+        expenseParsedFilters.value,
+      ],
       queryFn: ({ pageParam: next }) =>
-        getGroupExpenses(group?.id!, pageSize, expenseParsedFilters.value, next),
+        getGroupExpenses(
+          group?.id!,
+          pageSize,
+          expenseParsedFilters.value,
+          next
+        ),
       getNextPageParam: (lastPage) => lastPage?.next || undefined,
       initialPageParam: "",
     });
 
   const expenses = data?.pages.flatMap((p) => p.expenses);
 
-    useEffect(() => {
+  useEffect(() => {
     const expenseFilters = localStorage.getItem("expenseFilter");
     if (expenseFilters) {
-      expenseParsedFilters.value= JSON.parse(expenseFilters);
+      expenseParsedFilters.value = JSON.parse(expenseFilters);
     }
   }, []);
 
@@ -96,109 +112,125 @@ const Expenses = () => {
       ? totalSpent[userMemberId]?.[group.currency] ?? 0
       : 0;
 
+  const hasAnySearchParams =
+    expenseParsedFilters.value.before !== null ||
+    expenseParsedFilters.value.after !== null ||
+    expenseParsedFilters.value.freeText !== "" ||
+    (expenseParsedFilters.value.labels !== undefined &&
+      expenseParsedFilters.value.labels.length > 0) ||
+    (expenseParsedFilters.value.participantsIds !== undefined &&
+      expenseParsedFilters.value.participantsIds.length > 0) ||
+    (expenseParsedFilters.value.payersIds !== undefined &&
+      expenseParsedFilters.value.payersIds.length > 0);
+
+  
   return (
-    <StyledExpenses>
-      {!expenses || expenses.length === 0 ? (
+  <StyledExpenses>
+    {!expenses || expenses.length === 0 ? (
+      hasAnySearchParams ? (
+        <div className="noData">
+          <div className="msg">No expenses found. Have a go and refine your search! 🧐</div>
+          <FaMagnifyingGlass className="icon" />
+        </div>
+      ) : (
         <div className="noData">
           <div className="msg">There are currently no expenses</div>
           <CiReceipt className="icon" />
         </div>
-      ) : (
-        <>
-          {totalsAreFetching ? (
-            <div className="spinnerTotals">
-              <Spinner />
+      )
+    ) : (
+      <>
+        {totalsAreFetching ? (
+          <div className="spinnerTotals">
+            <Spinner />
+          </div>
+        ) : (
+          <BarsWithLegends
+            bar1Legend="Group Total"
+            bar2Legend="Your Share"
+            bar1Total={totalExpense || 0}
+            bar2Total={userExpense || 0}
+            currency={group.currency}
+            bar2Color="#e151ee"
+            bar1Color="#5183ee"
+            onClick={() => {
+              if (shouldOpenMultiCurrencyTable) {
+                menu.value = "epensesByCurrency";
+              }
+            }}
+          />
+        )}
+        {Object.entries(
+          groupBy(expenses, (x) => DateOnly(x.occurred, timeZoneId))
+        ).map(([date, expenses]) => (
+          <div key={date} className="same-date-container">
+            <div className="date-only">{date}</div>
+            <div className="expenses">
+              {expenses.map((e) => (
+                <div className="expense" key={e.id}>
+                  <Expense
+                    amount={e.amount}
+                    currency={e.currency}
+                    occurred={e.occurred}
+                    description={e.description}
+                    location={e.location}
+                    timeZoneId={timeZoneId}
+                    onClick={() => (selectedExpense.value = e)}
+                    userAmount={
+                      e.shares.find((x) => x.memberId === userMemberId)?.amount ?? 0
+                    }
+                    labels={e.labels}
+                  />
+                </div>
+              ))}
             </div>
-          ) : (
-            <BarsWithLegends
-              bar1Legend="Group Total"
-              bar2Legend="Your Share"
-              bar1Total={totalExpense || 0}
-              bar2Total={userExpense || 0}
-              currency={group.currency}
-              bar2Color="#e151ee"
-              bar1Color="#5183ee"
-              onClick={() => {
-                if (shouldOpenMultiCurrencyTable) {
-                  menu.value = "epensesByCurrency";
-                } else null;
-              }}
-            />
-          )}
-          {Object.entries(
-            groupBy(expenses, (x) => DateOnly(x.occurred, timeZoneId))
-          ).map(([date, expenses]) => (
-            <div key={date} className="same-date-container">
-              <div className="date-only">{date}</div>
-              <div className="expenses">
-                {expenses.map((e) => (
-                  <div className="expense" key={e.id}>
-                    <Expense
-                      amount={e.amount}
-                      currency={e.currency}
-                      occurred={e.occurred}
-                      description={e.description}
-                      location={e.location}
-                      timeZoneId={timeZoneId}
-                      onClick={() => (selectedExpense.value = e)}
-                      userAmount={
-                        e.shares.find((x) => x.memberId === userMemberId)
-                          ?.amount ?? 0
-                      }
-                      labels={e.labels}
-                    />
-                  </div>
-                ))}
-              </div>
-            </div>
-          ))}
-        </>
-      )}
-      <Sentinel
-        fetchNextPage={fetchNextPage}
-        hasNextPage={hasNextPage}
-        isFetchingNextPage={isFetchingNextPage}
+          </div>
+        ))}
+      </>
+    )}
+    <Sentinel
+      fetchNextPage={fetchNextPage}
+      hasNextPage={hasNextPage}
+      isFetchingNextPage={isFetchingNextPage}
+    />
+    {selectedExpense.value && (
+      <DetailedExpense
+        selectedExpense={selectedExpense}
+        amount={selectedExpense.value.amount}
+        currency={selectedExpense.value.currency}
+        description={selectedExpense.value.description}
+        labels={selectedExpense.value.labels}
+        location={selectedExpense.value.location}
+        occurred={selectedExpense.value.occurred}
+        payments={selectedExpense.value.payments}
+        shares={selectedExpense.value.shares}
+        timeZoneId={timeZoneId}
+        timeZoneCoordinates={userInfo.timeZoneCoordinates}
+        creator={selectedExpense.value.creatorId}
+        created={selectedExpense.value.created}
+        members={allParticipants}
+        errorMessage={errorMessage}
+        userMemberId={userMemberId || ""}
+        group={group}
       />
-      {selectedExpense.value && (
-        <DetailedExpense
-          selectedExpense={selectedExpense}
-          amount={selectedExpense.value.amount}
-          currency={selectedExpense.value.currency}
-          description={selectedExpense.value.description}
-          labels={selectedExpense.value.labels}
-          location={selectedExpense.value.location}
-          occurred={selectedExpense.value.occurred}
-          payments={selectedExpense.value.payments}
-          shares={selectedExpense.value.shares}
-          timeZoneId={timeZoneId}
-          timeZoneCoordinates={userInfo.timeZoneCoordinates}
-          creator={selectedExpense.value.creatorId}
-          created={selectedExpense.value.created}
-          members={allParticipants}
-          errorMessage={errorMessage}
-          userMemberId={userMemberId || ""}
-          group={group}
-        />
-      )}
-
-      <MenuAnimationBackground menu={menu} />
-
-      <ErrorMenuAnimation
-        menu={menu}
-        message={errorMessage.value}
-        type="expense"
-      />
-      <GroupTotalsByCurrencyAnimation
-        menu={menu}
-        bar1Legend="Group Total"
-        bar2Legend="Your Share"
-        bar2Color="#e151ee"
-        bar1Color="#5183ee"
-        groupTotalsByCurrency={groupTotalsByCurrency}
-        userTotalsByCurrency={userTotalsByCurrency}
-      />
-    </StyledExpenses>
-  );
+    )}
+    <MenuAnimationBackground menu={menu} />
+    <ErrorMenuAnimation
+      menu={menu}
+      message={errorMessage.value}
+      type="expense"
+    />
+    <GroupTotalsByCurrencyAnimation
+      menu={menu}
+      bar1Legend="Group Total"
+      bar2Legend="Your Share"
+      bar2Color="#e151ee"
+      bar1Color="#5183ee"
+      groupTotalsByCurrency={groupTotalsByCurrency}
+      userTotalsByCurrency={userTotalsByCurrency}
+    />
+  </StyledExpenses>
+);
 };
 
 export default Expenses;
